fix wrong opengl datatype

+++ NextEngine/Animation.cpp
@@ -30,7 +30,7 @@ void Animation::render(glm::mat4 globalModelTransform, Transform& transform, con
 	GLuint scaleXId = GameEngine::getInstance()->getRenderer()->getScaleXUniformId();
 	GLuint scaleYId = GameEngine::getInstance()->getRenderer()->getScaleYUniformId();
 	GLint colorOverlayId = GameEngine::getInstance()->getRenderer()->getColorOverlayUniformId();
-	GLfloat brightnessId = GameEngine::getInstance()->getRenderer()->getBrightnessUniformId();
+	GLuint brightnessId = GameEngine::getInstance()->getRenderer()->getBrightnessUniformId();
 
 	if (modelMatixId == -1) {
 		cout << "Error: Can't perform transformation " << endl;
